+++ config/shortcodes.py
