+++ include/meta_shortcodes.py
@@ -1,0 +1,4 @@
+import config.config as config
+def MetaShortCodes(data):
+	data = data.replace("[[URL]]",config.url+"/")
+	return data
